chore: move micro session and learning element creation to corresponding services
