+++ packages/graphql/src/services/microLearning.ts
@@ -1,6 +1,8 @@
-import { QuestionType } from '@klicker-uzh/prisma'
+import { Attachment, Question, QuestionType } from '@klicker-uzh/prisma'
+import { GraphQLError } from 'graphql'
 import { pick } from 'ramda'
 import { ContextWithUser } from '../lib/context'
+import { prepareInitialInstanceResults, processQuestionData } from './sessions'
 
 interface GetMicroSessionDataArgs {
   id: string
@@ -87,4 +89,95 @@ export async function markMicroSessionCompleted(
       },
     },
   })
+}
+
+interface CreateMicroSessionArgs {
+  name: string
+  displayName: string
+  description?: string
+  questions: number[]
+  courseId?: string
+  multiplier: number
+  startDate: Date
+  endDate: Date
+}
+
+export async function createMicroSession(
+  {
+    name,
+    displayName,
+    description,
+    questions,
+    courseId,
+    multiplier,
+    startDate,
+    endDate,
+  }: CreateMicroSessionArgs,
+  ctx: ContextWithUser
+) {
+  const dbQuestions = await ctx.prisma.question.findMany({
+    where: {
+      id: { in: questions },
+      ownerId: ctx.user.sub,
+    },
+    include: {
+      attachments: true,
+    },
+  })
+
+  const uniqueQuestions = new Set(dbQuestions.map((q) => q.id))
+  if (dbQuestions.length !== uniqueQuestions.size) {
+    throw new GraphQLError('Not all questions could be found')
+  }
+
+  const questionMap = dbQuestions.reduce<
+    Record<number, Question & { attachments: Attachment[] }>
+  >((acc, question) => ({ ...acc, [question.id]: question }), {})
+
+  const session = await ctx.prisma.microSession.create({
+    data: {
+      name,
+      displayName: displayName ?? name,
+      description,
+      pointsMultiplier: multiplier,
+      scheduledStartAt: new Date(startDate),
+      scheduledEndAt: new Date(endDate),
+      instances: {
+        create: questions.map((questionId, ix) => {
+          const question = questionMap[questionId]
+          const processedQuestionData = processQuestionData(question)
+          const questionAttachmentInstances = question.attachments.map(
+            pick(['type', 'href', 'name', 'description', 'originalName'])
+          )
+          return {
+            order: ix,
+            questionData: processedQuestionData,
+            results: prepareInitialInstanceResults(processedQuestionData),
+            question: {
+              connect: { id: questionId },
+            },
+            owner: {
+              connect: { id: ctx.user.sub },
+            },
+            attachments: {
+              create: questionAttachmentInstances,
+            },
+          }
+        }),
+      },
+      owner: {
+        connect: { id: ctx.user.sub },
+      },
+      course: {
+        connect: { id: courseId },
+      },
+    },
+    include: {
+      instances: true,
+    },
+  })
+
+  ctx.emitter.emit('invalidate', { typename: 'MicroSession', id: session.id })
+
+  return session
 }
