+++ packages/graphql/src/services/sessions.ts
@@ -15,11 +15,10 @@ import { ContextWithOptionalUser, ContextWithUser } from '../lib/context'
 import { GraphQLError } from 'graphql'
 import { max, mean, median, min, quantileSeq, std } from 'mathjs'
 import schedule from 'node-schedule'
-import { OrderType } from '../ops'
 
 const scheduledJobs: Record<string, any> = {}
 
-function processQuestionData(question: Question): AllQuestionTypeData {
+export function processQuestionData(question: Question): AllQuestionTypeData {
   switch (question.type) {
     case QuestionType.SC:
     case QuestionType.MC:
@@ -47,7 +46,9 @@ function processQuestionData(question: Question): AllQuestionTypeData {
   }
 }
 
-function prepareInitialInstanceResults(questionData: AllQuestionTypeData) {
+export function prepareInitialInstanceResults(
+  questionData: AllQuestionTypeData
+) {
   switch (questionData.type) {
     case QuestionType.SC:
     case QuestionType.MC:
@@ -337,191 +338,6 @@ export async function getLiveSessionData(
   })
 
   return session
-}
-
-interface CreateMicroSessionArgs {
-  name: string
-  displayName: string
-  description?: string
-  questions: number[]
-  courseId?: string
-  multiplier: number
-  startDate: Date
-  endDate: Date
-}
-
-export async function createMicroSession(
-  {
-    name,
-    displayName,
-    description,
-    questions,
-    courseId,
-    multiplier,
-    startDate,
-    endDate,
-  }: CreateMicroSessionArgs,
-  ctx: ContextWithUser
-) {
-  const dbQuestions = await ctx.prisma.question.findMany({
-    where: {
-      id: { in: questions },
-      ownerId: ctx.user.sub,
-    },
-    include: {
-      attachments: true,
-    },
-  })
-
-  const uniqueQuestions = new Set(dbQuestions.map((q) => q.id))
-  if (dbQuestions.length !== uniqueQuestions.size) {
-    throw new GraphQLError('Not all questions could be found')
-  }
-
-  const questionMap = dbQuestions.reduce<
-    Record<number, Question & { attachments: Attachment[] }>
-  >((acc, question) => ({ ...acc, [question.id]: question }), {})
-
-  const session = await ctx.prisma.microSession.create({
-    data: {
-      name,
-      displayName: displayName ?? name,
-      description,
-      pointsMultiplier: multiplier,
-      scheduledStartAt: new Date(startDate),
-      scheduledEndAt: new Date(endDate),
-      instances: {
-        create: questions.map((questionId, ix) => {
-          const question = questionMap[questionId]
-          const processedQuestionData = processQuestionData(question)
-          const questionAttachmentInstances = question.attachments.map(
-            pick(['type', 'href', 'name', 'description', 'originalName'])
-          )
-          return {
-            order: ix,
-            questionData: processedQuestionData,
-            results: prepareInitialInstanceResults(processedQuestionData),
-            question: {
-              connect: { id: questionId },
-            },
-            owner: {
-              connect: { id: ctx.user.sub },
-            },
-            attachments: {
-              create: questionAttachmentInstances,
-            },
-          }
-        }),
-      },
-      owner: {
-        connect: { id: ctx.user.sub },
-      },
-      course: {
-        connect: { id: courseId },
-      },
-    },
-    include: {
-      instances: true,
-    },
-  })
-
-  ctx.emitter.emit('invalidate', { typename: 'MicroSession', id: session.id })
-
-  return session
-}
-
-interface CreateLearningElementArgs {
-  name: string
-  displayName: string
-  description?: string
-  questions: number[]
-  courseId?: string
-  multiplier: number
-  order: OrderType
-  resetTimeDays: number
-}
-
-export async function createLearningElement(
-  {
-    name,
-    displayName,
-    description,
-    questions,
-    courseId,
-    multiplier,
-    order,
-    resetTimeDays,
-  }: CreateLearningElementArgs,
-  ctx: ContextWithUser
-) {
-  const dbQuestions = await ctx.prisma.question.findMany({
-    where: {
-      id: { in: questions },
-      ownerId: ctx.user.sub,
-    },
-    include: {
-      attachments: true,
-    },
-  })
-
-  const uniqueQuestions = new Set(dbQuestions.map((q) => q.id))
-  if (dbQuestions.length !== uniqueQuestions.size) {
-    throw new GraphQLError('Not all questions could be found')
-  }
-
-  const questionMap = dbQuestions.reduce<
-    Record<number, Question & { attachments: Attachment[] }>
-  >((acc, question) => ({ ...acc, [question.id]: question }), {})
-
-  const element = await ctx.prisma.learningElement.create({
-    data: {
-      name,
-      displayName: displayName ?? name,
-      description,
-      pointsMultiplier: multiplier,
-      orderType: order,
-      resetTimeDays: resetTimeDays,
-      instances: {
-        create: questions.map((questionId, ix) => {
-          const question = questionMap[questionId]
-          const processedQuestionData = processQuestionData(question)
-          const questionAttachmentInstances = question.attachments.map(
-            pick(['type', 'href', 'name', 'description', 'originalName'])
-          )
-          return {
-            order: ix,
-            questionData: processedQuestionData,
-            results: prepareInitialInstanceResults(processedQuestionData),
-            question: {
-              connect: { id: questionId },
-            },
-            owner: {
-              connect: { id: ctx.user.sub },
-            },
-            attachments: {
-              create: questionAttachmentInstances,
-            },
-          }
-        }),
-      },
-      owner: {
-        connect: { id: ctx.user.sub },
-      },
-      course: {
-        connect: { id: courseId },
-      },
-    },
-    include: {
-      instances: true,
-    },
-  })
-
-  ctx.emitter.emit('invalidate', {
-    typename: 'LearningElement',
-    id: element.id,
-  })
-
-  return element
 }
 
 interface StartSessionArgs {
